+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added cmdlet `Set-PSModulePath`
+- Added cmdlet `Set-PSModulePath`.
 
 ## [0.8.0] - 2020-05-11